fix(test): replace no-op localStorage stub with working in-memory implementation
The global test setup stubbed localStorage with bare vi.fn()s that discarded
all writes and returned undefined (not null) for missing keys. This meant any
test that wrote a value then read it back was asserting against a mock that
could not round-trip data — tests either passed vacuously or failed for the
wrong reason.

Replace the stub with a Map-backed in-memory Storage that matches the real
Web Storage API (getItem returns null for absent keys, setItem persists,
clear/removeItem/key/length all work). A fresh instance is installed after
each test so state never leaks.

With the global mock now functional:
- Remove installMemoryStorage() workaround from network-weather.test.ts
- Remove file-local localStorage mock from posting-time-recommendations.test.ts

All 1345 tests pass (69 files), build clean, format clean.

[1217171056973042]
